+++ lib/core/models/shuttle_arrival.dart
