+++ lib/core/models/shuttle_stop.dart
