+++ lib/core/models/spot_types.dart
