+++ lib/core/services/spot_types.dart
@@ -1,12 +1,11 @@
 import 'package:campus_mobile_experimental/app_networking.dart';
-import 'package:campus_mobile_experimental/core/models/parking_model.dart';
+import 'package:campus_mobile_experimental/core/models/spot_types.dart';
 
-class ParkingService {
-  ParkingService() {
-    fetchParkingLotData();
+class SpotTypesService {
+  SpotTypesService() {
+    fetchSpotTypesData();
   }
   bool _isLoading = false;
-  List<ParkingModel> _data;
   DateTime _lastUpdated;
   String _error;
   final NetworkHelper _networkHelper = NetworkHelper();
@@ -14,18 +13,18 @@ class ParkingService {
     "accept": "application/json",
   };
 
-  final String endpoint =
-      "https://mobile.ucsd.edu/replatform/v1/qa/webview/parking-v2/parking_lots.json";
+  SpotTypeModel _spotTypeModel = SpotTypeModel();
 
-  Future<bool> fetchParkingLotData() async {
+  final String endpoint =
+      "https://mobile.ucsd.edu/replatform/v1/qa/webview/parking-v2/spot_types.json";
+  Future<bool> fetchSpotTypesData() async {
     _error = null;
     _isLoading = true;
     try {
       /// fetch data
       String _response = await _networkHelper.fetchData(endpoint);
+      _spotTypeModel = spotTypeModelFromJson(_response);
 
-      /// parse data
-      _data = parkingModelFromJson(_response);
       _isLoading = false;
       return true;
     } catch (e) {
@@ -35,8 +34,8 @@ class ParkingService {
     }
   }
 
-  List<ParkingModel> get data => _data;
   bool get isLoading => _isLoading;
   String get error => _error;
   DateTime get lastUpdated => _lastUpdated;
+  SpotTypeModel get spotTypeModel => _spotTypeModel;
 }
