+++ lib/ui/scanner/scanner_view.dart
@@ -67,11 +67,9 @@ class _ScanditScannerState extends State<ScanditScanner> {
   Widget build(BuildContext context) {
     _userDataProvider = Provider.of<UserDataProvider>(context);
     if (Theme.of(context).platform == TargetPlatform.iOS) {
-      licenseKey =
-          "AWZPNIuBH8uOAj5WPTefVbAXwJZoCGikFkOOoRJADibsLbKyxEmxtKtde2Vbcn+tt3dhyaxjwO05RsNOJG6qEFFnsNTWRCpbJBStZctoZG6UYhICzXWawCx0EmmULQQy5xwn1rwN7KPROh5hj/QFyQyskdZs7ltbOAfSMfvTe63o5nARzGe2nPgz3C6KGgJraGyK9BB3vs0Gg2JDpJk1MmFAAs8XE59UTxGmCHfOTnY8GEOjtGrcy52HnjsqDa++XGbJHqHxQ350NZDqbi3X+zd7OQNGZ6UjarkdsXODx7mNOzfgDtVzvyF7TOJsyWvyauP2QSXutnShkZP0+BqJiK3KAfrhRSK3BTN/9CHNAcWBRvxFzkqAc3zvkXJazRxRBDIBxV+gYESCbK9/mv9eB93M79uVTlDrTGXFj3iJbZ8fuqEhXQivQy8yGnqoxaVcaBQsltn2uOa+lzsCGuprBjrn8MlBTItcStidxAEd5rhDQ6h37/+ZeYLBc8Mz/Qvd+LHfY2tJricL+x7UuwBq/wh2/f6mwIKvjeVNR6ij1ugJPchKapCG7PXxosByKW9Iz/NF8+ItrIxlTHQwSj0N3xtlA26BKosHyl7r4Se5N2RB9WQWL7Ql5IDR2kzVg7dmKG1NWimZbCSA9P6XwWDJ2iSLMSsiEofQPM9upIzkW0xlwe1JqvtXJtXkkaEmFsNAN/9nrMSzmiiEWUHF23IFtLUMX0uXaw/5voG6wmzUzKza/pwpWzt6BWwKE9OOiVK59nNMMlOLPLHe8MOeBVdn7CSUugVcx9X2b53mQX/NE//dQmFG/377G1iK";
+      licenseKey = "SCANDIT_NATIVE_LICENSE_IOS_PH";
     } else if (Theme.of(context).platform == TargetPlatform.android) {
-      licenseKey =
-          "AcSv56yBOHYQPJDNviC89nU41XcdNK3kmEqQvPtTx6usVyIvbmkbs8d0dz2ecugTsnaqBdwvK17sZghABykr/JpQ1IP8VKA+cAoZCrxwpo70HKTDJydnIrw17B9zCb1kO2qm1jlfCCq+HMRcEIB/ZrHDu/hNCilVw7aOo6+lqzOqSjVaek8rKUZZnsQ6SQnDYi4+XG2n5r9RVp5DZogkduBMGr2U2ZG73Uj/8I6ZgvQAwiMOkP0rDP7oME7KwSrzjZVQbOgb7JC3PvrHPsxq0CmaXCsOBHpQ2Ri0JRUgvSt78JBbObmqhj7pHSlrCvkzIkxDSGi+s1x0hNktVuaUXcqcEIu3GpcDmEER++QUX56zjn3BF0nM5ecjx+03DyRjEJ79xmIvUn/9kdOTWCVfQOr2Vu9ElJ02uaGr0l29NZW39T5ZkYjSqzwt+qnV3AT+GIk+r7jH62pcJK/GXfSeQNYVNwdUVr/iJSbwWVFGvhrpgQds/hoB7dNlQ/YpUSffE5aogL/Idlp6nH3t364p2enRAh7B3uJx+e/DXck72J8BycSJrl4N4J8F0AFaXI3opppIRc9/fXaLOunUrJcMULDRhkkzXRFoKszPJjqtRNPe/1+VFgwbS0LXB/zFt9FM4E04RC/ZHPA3HwT/PbpnoPdPoarCD1M4IfwnEsXVGzVSj6Q+TZHubazSkQszQX/qIhpxwW0oGJRYSwSXygwhqfrqvi4fYOmRSox1mnngyOfLtJMZU29i/mcpO1Ib7Tcyer90V0GvnkyQ0wq9hFUOhS6nVRIJXcX68mYIQDDo";
+      licenseKey = "SCANDIT_NATIVE_LICENSE_ANDROID_PH";
     }
     return Scaffold(
       appBar: PreferredSize(
