+++ lib/ui/shuttle/shuttle_display.dart
@@ -1,6 +1,6 @@
-import 'package:campus_mobile_experimental/core/models/shuttle_arrival_model.dart';
-import 'package:campus_mobile_experimental/core/models/shuttle_model.dart';
-import 'package:campus_mobile_experimental/core/models/shuttle_stop_model.dart';
+import 'package:campus_mobile_experimental/core/models/shuttle.dart';
+import 'package:campus_mobile_experimental/core/models/shuttle_arrival.dart';
+import 'package:campus_mobile_experimental/core/models/shuttle_stop.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
